Move type tests to type tests folder
