+++ ember-resources/src/index.ts
@@ -4,7 +4,7 @@ export { resource, resourceFactory } from './core/function-based';
 export { use } from './core/use';
 
 // Public API -- Utilities
-export { cell } from 'util/cell';
+export { cell } from './util/cell';
 
 // Public Type Utilities
 export type { Hooks as ResourceAPI } from './core/function-based';
