+++ ember-resources/type-tests/core-types.test.ts
