+++ ember-resources/type-tests/thunk-helpers.test.ts
